pyproject.toml
[tool.poetry]
authors = ["FooQoo <horseshoes.buzz.17@gmail.com>"]
description = "The batch system for clawling stocks"
license = "MIT"
name = "stock-crawler"
version = "1.0.0"

[tool.poetry.dependencies]
google-api-python-client = "^1.10.0"
jupyter = "^1.0.0"
logger = "^1.4"
matplotlib = "^3.3.0"
oauth2client = "^4.1.3"
pandas = "^1.0.5"
python = "^3.8"
python-binary-memcached = "^0.29.0"
python-dotenv = "^0.14.0"
requests = "pandas"
seaborn = "^0.10.1"
tqdm = "^4.48.0"
google-cloud-storage = "^1.31.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.3"
flake8 = "^3.8.3"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
